Add annotations to instantiate beans
The FakePersonDataAccessService class needs to be instantiated as a
bean so it can be injected into other classes (dependency injection). We
do this by adding annotations to both FakePersonDataAccessService and
the PersonService class.

I'm not really sure how dependency injection works through and through
but I'll continue to follow along in the hopes that it will become more
clear later in the tutorial.