Add @RestController annotation to PersonController
PersonController is not available as a RestController so we need to add
the @RestController annotation. We can then have some methods, and then
expose some endpoints that clients can consume.

Going back to our architecture, we can note that we would be receiving
HTTP requests (GET, POST, DELETE, and PUT). For whatever client we have
(React, Android, etc.) to issue requests to our backend, we need to make
the class available as  RestController.