Add 2 new methods (update & delete) to PersonDao
We now want to add the ability to update a person, as well as deleting
a person. For this, we start by adding some bodyless methods to our
PersonDao interface (data access object).

These methods are:

1) updatePersonByID -> Takes an id and presumably the new Person object
we
are updating to.

2) deletePersonById -> Takes an id.

3) selectPersonById -> We'll probably use this as a utility method for
the others if I had to guess. We are using Optional probably because we
may or may not have some null values, I'm not really sure yet; I'll have
to keep following the tutorial.